+++ app.py
@@ -1,22 +1,26 @@
-import os
-import json
 import asyncio
-from pathlib import Path
+import os
 from datetime import datetime
-from urllib.request import urlopen, Request
-from typing import Union
 from collections import Counter
-from configparser import ConfigParser
 from random import randint
+from urllib.request import urlopen, Request
 
-import ffmpeg
 import discord
 from discord.ext.commands import Context, Bot
 from discord import FFmpegOpusAudio, Embed, Intents
 from openai import AsyncOpenAI
-from PIL import Image
 
-from db_utils import get_thread_id
+from ai_helpers import (
+    nonsense_talk,
+    gs_intro_song,
+    would_you_rather,
+    get_config,
+    get_trivia_question,
+    dict_to_ordered_string,
+    generate_speech,
+    content_path,
+    is_square_image,
+)
 
 
 # OpenAI Client
@@ -51,13 +55,12 @@ async def join(ctx: Context, arg1: int = 0, arg2: int = 5):
 
         while True:
 
-            await asyncio.sleep(interval)
-
-            tts, path = await nonsense_talk(ctx.guild.id)
-            source = FFmpegOpusAudio(path)
+            tts, file_path = await nonsense_talk(ctx.guild.id)
+            source = FFmpegOpusAudio(file_path)
             player = voice.play(source)
 
             await ctx.send(tts)
+            await asyncio.sleep(interval)
 
 
 @bot.command()
@@ -74,8 +77,8 @@ async def theme(ctx: Context, arg1: str):
     """
 
     voice = discord.utils.get(bot.voice_clients, guild=ctx.guild)
-    tts, path = await gs_intro_song(ctx.guild.id, arg1)
-    source = FFmpegOpusAudio(path)
+    tts, file_path = await gs_intro_song(ctx.guild.id, arg1)
+    source = FFmpegOpusAudio(file_path)
     player = voice.play(source)
     await ctx.send(tts)
 
@@ -89,8 +92,8 @@ async def rather(ctx: Context, arg1: str = "normal"):
 
     arg1 = arg1.lower()
     voice = discord.utils.get(bot.voice_clients, guild=ctx.guild)
-    tts, path = await would_you_rather(guild_id=ctx.guild.id, topic=arg1)
-    source = FFmpegOpusAudio(path)
+    tts, file_path = await would_you_rather(guild_id=ctx.guild.id, topic=arg1)
+    source = FFmpegOpusAudio(file_path)
     player = voice.play(source)
     await ctx.send(tts)
 
@@ -393,153 +396,6 @@ async def edit(ctx: Context, *, arg: str = ""):
     file_upload = discord.File(path, filename=file_name)
 
     await ctx.send(file=file_upload, embed=embed)
-
-
-async def new_thread_response(
-    thread_name: str, prompt: str = "", guild_id: str = "", response_format: Union[str, dict] = "auto"
-):
-
-    config = get_config()
-    assistant_id = config.get("OPENAI_ASSISTANTS", thread_name)
-
-    thread_id = await get_thread_id(guild_id=guild_id, name=thread_name, assistant_id=assistant_id, client=client)
-
-    # add a message to the thread
-    await client.beta.threads.messages.create(
-        thread_id=thread_id,
-        role="user",
-        content=prompt,
-    )
-
-    # run the thread with the assistant and monitor the situation
-    run = await client.beta.threads.runs.create_and_poll(
-        thread_id=thread_id, assistant_id=assistant_id, response_format=response_format
-    )
-
-    messages = await client.beta.threads.messages.list(thread_id=thread_id)
-
-    # get the most recent response from the assistant
-    response = messages.data[0]
-
-    return response
-
-
-async def generate_speech(guild_id: str, compartment: str, file_name: str, tts: str) -> str:
-    config = get_config()
-    async with client.audio.speech.with_streaming_response.create(
-        model=config.get("OPENAI_GENERAL", "speech_model", fallback="tts-1"),
-        voice=config.get("OPENAI_GENERAL", "voice", fallback="onyx"),
-        input=tts,
-        response_format=config.get("OPENAI_GENERAL", "speech_file_format", fallback="wav"),
-    ) as speech:
-        file_path = content_path(guild_id=guild_id, compartment=compartment, file_name=file_name)
-        await speech.stream_to_file(file_path)
-
-    return file_path
-
-
-async def gs_intro_song(guild_id: str, name: str):
-    config = get_config()
-    name = f"{name}_theme"
-
-    prompt = config.get("PROMPTS", name)
-
-    # openai api work
-    ## generate the show intro text
-    response = await new_thread_response(prompt=prompt, guild_id=guild_id, thread_name="gs_host")
-
-    ## generate a speech wav
-    tts = response.content[0].text.value
-    file_path = await generate_speech(
-        guild_id=guild_id, compartment="theme", tts=tts, file_name=f"{name}_{response.id}.wav"
-    )
-
-    # ffmpeg work to combine streams
-    ## load both audio files
-    theme_song = ffmpeg.input("gameshow.mp3").audio
-    words = ffmpeg.input(file_path).audio
-
-    ## adjust volumes
-    theme_song = theme_song.filter("volume", 0.25)
-    words = words.filter("volume", 3)
-
-    ## merge and output
-    dt_string = datetime.now().strftime("%Y%m%d%H%M%S")
-    ouput_file = Path(f"intro_{dt_string}.wav").resolve()
-    merged_audio = ffmpeg.filter((theme_song, words), "amix")
-    out = ffmpeg.output(merged_audio, str(ouput_file)).overwrite_output()
-    out.run(quiet=True)
-
-    return tts, ouput_file
-
-
-async def would_you_rather(guild_id: str, topic: str):
-    config = get_config()
-    thread_name = f"rather_{topic}"
-
-    new_hypothetical_prompt = config.get("PROMPTS", "new_hypothetical")
-
-    response = await new_thread_response(thread_name=thread_name, prompt=new_hypothetical_prompt, guild_id=guild_id)
-    tts = response.content[0].text.value
-    file_path = await generate_speech(guild_id=guild_id, compartment="rather", tts=tts, file_name=f"{response.id}.wav")
-
-    return tts, file_path
-
-
-async def nonsense_talk(guild_id: str):
-
-    response = await new_thread_response(thread_name="nonsense_talk", prompt="Let's hear it.", guild_id=guild_id)
-    tts = response.content[0].text.value
-    file_path = await generate_speech(guild_id=guild_id, compartment="rather", tts=tts, file_name=f"{response.id}.wav")
-
-    return tts, file_path
-
-
-async def get_trivia_question(guild_id: str) -> dict:
-    config = get_config()
-    trivia_prompt = config.get("PROMPTS", "trivia_game")
-
-    response = await new_thread_response(
-        thread_name="trivia_game", prompt=trivia_prompt, guild_id=guild_id, response_format={"type": "json_object"}
-    )
-
-    text_json = response.content[0].text.value
-
-    response_dict = json.loads(text_json)
-
-    # may be something to do, here.
-
-    return response_dict
-
-
-def content_path(guild_id: str, compartment: str, file_name: str):
-    config = get_config()
-    ts = datetime.now().strftime(config.get("GENERAL", "session_strftime", fallback="dall-e-2"))
-    dir_path = Path(f"generated_content/guild_{guild_id}/{ts}/{compartment}")
-    dir_path.mkdir(parents=True, exist_ok=True)
-    return dir_path / file_name
-
-
-def is_square_image(image_path: Path):
-    with Image.open(image_path) as img:
-        width, height = img.size
-        return width == height
-
-
-def dict_to_ordered_string(data: dict) -> str:
-    # Sort the dictionary items by value in descending order
-    sorted_items = sorted(data.items(), key=lambda item: item[1], reverse=True)
-
-    # Format the sorted items into a numbered list
-    formatted_string = "\n".join([f"{i+1}. **{key}**: {value}" for i, (key, value) in enumerate(sorted_items)])
-
-    return formatted_string
-
-
-def get_config():
-    config = ConfigParser()
-    config.read("config.ini")
-    return config
 
 
 bot.run(os.getenv("DISCORD_BOT_KEY"))
