+++ config.ini
@@ -38,4 +38,5 @@ new_hypothetical = """Ask me a new hypothetical question. The question should re
     Make sure it is completely unlike every other hypothetical question in this thread.
     The question should start an interesting conversation in a chat room.
     """
-trivia_game = "Can I have a new question unlike any of the others in this thread?"
+trivia_game = Can I have a new question unlike any of the others in this thread?
+nonsense_talk = Generate new yelling words.
